Fix failing regression tests by reverting changes to parse_utf16.
There appears to be a bug in the handling of escaped utf16 characters. The bug
looks like a problem with an uninitialized variable in parse_utf16.  It was
introduced as part of the conversion and was causing regression tests to fail.

I found the conversion changes in parse_utf16 complicated to follow, so I
reverted them.  I had to introduce some temporary unchecked pointer variables
at the call to parse_utf16 due to restrictions on taking the address of local
variables with bounds.