Fix compiler warnings about bounds declarations, clean up code.
There were 17 warnings about being unable to prove bounds declarations.
Change the source code for parson.c to reduce the number of warnings to 6,
5 of which are instances  of Github issue #599.  Also clean up the code by
removing unneeded changes.

1. Introduce a helper function for allocating strings.  This can't
be done in checked code because it requires casting to Nt_array_ptr,
which isn't allowed by the specification (this restriction is not
implemented by the compiler yet).
2. parson_strndup doesn't need to have a return bounds of count(n).
While this is technically correct for this implementation of strndup,
it might be wrong for other implementations of strndup, which may not
actually create a string of size n given a string with fewer than
n character.
3. process_string specified the bounds for input_ptr and output_ptr
incorrectly.  It used counts, but then adjusted input_ptr and output_ptr,
which leads to be more complicated book keeping for bounds (which was
missing).  Just  use a range.
4. Remove bounds declarations on dot_position local variables in
json_object_dotget_value, json_object_dotset_value, and json_object_dotremove.
They aren't needed.  It's fine to let dot_position just be a string with
default bounds of count(0).

Clean up:
1. read_file doesn't need to calculate its length.  It's fine for it to just
return a null-terminated string.
2. Undo changes in json_parse_file and json_parse_file_with_comments
to allow calls to read_file to calculate lengths.  The length isn't
needed.
3. Revert changes to parson_strdup; they're  not needed.
4. Delete out-of-date comment about an odd compiler error in
json_serialize_string.  With the corrected bounds declarations from the prior
check-in, this no  longer happens.